# generated/aspnetcore/src/API/Models/Bond.cs
/*
 * Financial Instrument Sample Polymorph
 *
 * This is the description.
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: openwealth@synpulse.com
 * Generated by: https://openapi-generator.tech
 */

using System;
using System.Linq;
using System.Text;
using System.Collections.Generic;
using System.ComponentModel;
using System.ComponentModel.DataAnnotations;
using System.Runtime.Serialization;
using System.Text.Json;
using API.Converters;

namespace API.Models
{ 
    /// <summary>
    /// Schema of a bond instrument.
    /// </summary>
    [DataContract]
    public partial class Bond : FinancialInstrument
    {
        /// <summary>
        /// currency of denomination.
        /// </summary>
        /// <value>currency of denomination.</value>
        [RegularExpression("[A-Z]{3}")]
        [DataMember(Name="currencyOfDenomination", EmitDefaultValue=false)]
        public string? CurrencyOfDenomination { get; set; }

        /// <summary>
        /// Maturity date of instrument.
        /// </summary>
        /// <value>Maturity date of instrument.</value>
        [DataMember(Name="maturityDate", EmitDefaultValue=true)]
        public DateOnly? MaturityDate { get; set; }

        /// <summary>
        /// Interest rate of the bond.
        /// </summary>
        /// <value>Interest rate of the bond.</value>
        [DataMember(Name="interestRate", EmitDefaultValue=true)]
        public decimal? InterestRate { get; set; }

    }
}