# generated/aspnetcore/src/API/Models/Cash.cs
/*
 * Financial Instrument Sample Polymorph
 *
 * This is the description.
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: openwealth@synpulse.com
 * Generated by: https://openapi-generator.tech
 */

using System;
using System.Linq;
using System.Text;
using System.Collections.Generic;
using System.ComponentModel;
using System.ComponentModel.DataAnnotations;
using System.Runtime.Serialization;
using System.Text.Json;
using API.Converters;

namespace API.Models
{ 
    /// <summary>
    /// Schema of a cash instrument.
    /// </summary>
    [DataContract]
    public partial class Cash : FinancialInstrument
    {
        /// <summary>
        /// Currency code.
        /// </summary>
        /// <value>Currency code.</value>
        [Required]
        [RegularExpression("^[A-Z]{3}$")]
        [DataMember(Name="currency", EmitDefaultValue=false)]
        public string Currency { get; set; }

    }
}