# generated/aspnetcore/src/API/Models/FinancialInstrument.cs
/*
 * Financial Instrument Sample Polymorph
 *
 * This is the description.
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: openwealth@synpulse.com
 * Generated by: https://openapi-generator.tech
 */

using System;
using System.Linq;
using System.Text;
using System.Collections.Generic;
using System.ComponentModel;
using System.ComponentModel.DataAnnotations;
using System.Runtime.Serialization;
using System.Text.Json;
using JsonSubTypes;
using Swashbuckle.AspNetCore.Annotations;
using API.Converters;

namespace API.Models
{ 
    /// <summary>
    /// Financial instrument entity.
    /// </summary>
    [DataContract]
    [SwaggerDiscriminator("Type")]
    [JsonSubtypes.KnownSubType(typeof(Bond), "Bond")]
    [SwaggerSubType(typeof(Bond), DiscriminatorValue =  "Bond")]
    [JsonSubtypes.KnownSubType(typeof(Cash), "Cash")]
    [SwaggerSubType(typeof(Cash), DiscriminatorValue =  "Cash")]
    [JsonSubtypes.KnownSubType(typeof(Equity), "Equity")]
    [SwaggerSubType(typeof(Equity), DiscriminatorValue =  "Equity")]
    [JsonSubtypes.KnownSubType(typeof(Option), "Option")]
    [SwaggerSubType(typeof(Option), DiscriminatorValue =  "Option")]
    public partial class FinancialInstrument 
    {

        /// <summary>
        /// Type of the financial instrument.
        /// </summary>
        /// <value>Type of the financial instrument.</value>
        
        public enum TypeEnum
        {
            
            /// <summary>
            /// Enum CashEnum for Cash
            /// </summary>
            [EnumMember(Value = "Cash")]
            CashEnum = 1,
            
            /// <summary>
            /// Enum BondEnum for Bond
            /// </summary>
            [EnumMember(Value = "Bond")]
            BondEnum = 2,
            
            /// <summary>
            /// Enum EquityEnum for Equity
            /// </summary>
            [EnumMember(Value = "Equity")]
            EquityEnum = 3,
            
            /// <summary>
            /// Enum OptionEnum for Option
            /// </summary>
            [EnumMember(Value = "Option")]
            OptionEnum = 4
        }

        /// <summary>
        /// Type of the financial instrument.
        /// </summary>
        /// <value>Type of the financial instrument.</value>
        [Required]
        [DataMember(Name="type", EmitDefaultValue=true)]
        public TypeEnum Type { get; set; }

        /// <summary>
        /// Name of the financial instrument.
        /// </summary>
        /// <value>Name of the financial instrument.</value>
        [DataMember(Name="name", EmitDefaultValue=false)]
        public string? Name { get; set; }

        /// <summary>
        /// List of identification key value pairs.
        /// </summary>
        /// <value>List of identification key value pairs.</value>
        [DataMember(Name="identificationList", EmitDefaultValue=false)]
        public List<Identification> IdentificationList { get; set; }

    }
}