# generated/aspnetcore/src/API/Models/Identification.cs
/*
 * Financial Instrument Sample Polymorph
 *
 * This is the description.
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: openwealth@synpulse.com
 * Generated by: https://openapi-generator.tech
 */

using System;
using System.Linq;
using System.Text;
using System.Collections.Generic;
using System.ComponentModel;
using System.ComponentModel.DataAnnotations;
using System.Runtime.Serialization;
using System.Text.Json;
using API.Converters;

namespace API.Models
{ 
    /// <summary>
    /// Schema of and identification key value pair.
    /// </summary>
    [DataContract]
    public partial class Identification 
    {

        /// <summary>
        /// Identification type.
        /// </summary>
        /// <value>Identification type.</value>
        
        public enum TypeEnum
        {
            
            /// <summary>
            /// Enum Iso3Enum for iso3
            /// </summary>
            [EnumMember(Value = "iso3")]
            Iso3Enum = 1,
            
            /// <summary>
            /// Enum IsinEnum for isin
            /// </summary>
            [EnumMember(Value = "isin")]
            IsinEnum = 2,
            
            /// <summary>
            /// Enum OtherEnum for other
            /// </summary>
            [EnumMember(Value = "other")]
            OtherEnum = 3
        }

        /// <summary>
        /// Identification type.
        /// </summary>
        /// <value>Identification type.</value>
        [Required]
        [DataMember(Name="type", EmitDefaultValue=true)]
        public TypeEnum Type { get; set; }

        /// <summary>
        /// Value of ientifier.
        /// </summary>
        /// <value>Value of ientifier.</value>
        [Required]
        [DataMember(Name="identifier", EmitDefaultValue=false)]
        public string Identifier { get; set; }

    }
}