# generated/aspnetcore/src/API/Models/Option.cs
/*
 * Financial Instrument Sample Polymorph
 *
 * This is the description.
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: openwealth@synpulse.com
 * Generated by: https://openapi-generator.tech
 */

using System;
using System.Linq;
using System.Text;
using System.Collections.Generic;
using System.ComponentModel;
using System.ComponentModel.DataAnnotations;
using System.Runtime.Serialization;
using System.Text.Json;
using API.Converters;

namespace API.Models
{ 
    /// <summary>
    /// Schema of a equity instrument.
    /// </summary>
    [DataContract]
    public partial class Option : FinancialInstrument
    {
        /// <summary>
        /// Gets or Sets UnderlyingInstrument
        /// </summary>
        [DataMember(Name="underlyingInstrument", EmitDefaultValue=false)]
        public FinancialInstrument? UnderlyingInstrument { get; set; }

    }
}